requirements.txt
Flask
Flask-Login
Flask-WTF
psycopg2-binary
Werkzeug
python-dotenv
